+++ finagle-memcached/src/test/scala/com/twitter/finagle/memcached/integration/MemcachedOldClientTest.scala
@@ -13,16 +13,15 @@ import java.net.InetSocketAddress
 
 class MemcachedOldClientTest extends MemcachedTest {
 
-  protected[this] val isOldClient: Boolean = true
-
   protected def createClient(dest: Name, clientName: String): Client = {
     Memcached.client.newRichClient(dest, clientName)
   }
 
-  protected[this] val redistributesKey: Seq[String] = Seq(clientName, "redistributes")
-  protected[this] val leavesKey: Seq[String] = Seq(clientName, "leaves")
-  protected[this] val revivalsKey: Seq[String] = Seq(clientName, "revivals")
-  protected[this] val ejectionsKey: Seq[String] = Seq(clientName, "ejections")
+  protected[this] val redistributesKey: Seq[String] =
+    Seq(clientName, "partitioner", "redistributes")
+  protected[this] val leavesKey: Seq[String] = Seq(clientName, "partitioner", "leaves")
+  protected[this] val revivalsKey: Seq[String] = Seq(clientName, "partitioner", "revivals")
+  protected[this] val ejectionsKey: Seq[String] = Seq(clientName, "partitioner", "ejections")
 
   test("re-hash when a bad host is ejected") {
     val sr = new InMemoryStatsReceiver
@@ -30,7 +29,7 @@ class MemcachedOldClientTest extends MemcachedTest {
       .configured(FailureAccrualFactory.Param(1, () => 10.minutes))
       .configured(param.EjectFailedHost(true))
       .withStatsReceiver(sr)
-      .newRichClient(Name.bound(servers.map { s => Address(s.address) }: _*), clientName)
+      .newTwemcacheClient(Name.bound(servers.map { s => Address(s.address) }: _*), clientName)
     testRehashUponEject(client, sr)
     client.close()
   }
@@ -42,7 +41,7 @@ class MemcachedOldClientTest extends MemcachedTest {
         .configured(EjectFailedHost(true))
         .configured(Timer(timer))
         .configured(Stats(statsReceiver))
-        .newRichClient(
+        .newTwemcacheClient(
           Name.bound(Address(cacheServer.boundAddress.asInstanceOf[InetSocketAddress])),
           clientName
         ))
@@ -59,7 +58,7 @@ class MemcachedOldClientTest extends MemcachedTest {
     val client = Memcached.client
       .connectionsPerEndpoint(NumConnections)
       .withStatsReceiver(sr)
-      .newRichClient(Name.Bound.singleton(mutableAddrs), "test_client")
+      .newTwemcacheClient(Name.Bound.singleton(mutableAddrs), "test_client")
 
     testAddAndRemoveNodes(addrs, mutableAddrs, sr)
     client.close()